+++ src/main/scala/edu/ie3/simona/config/ConfigFailFast.scala
@@ -9,7 +9,12 @@ package edu.ie3.simona.config
 import com.typesafe.config.{Config, ConfigException}
 import com.typesafe.scalalogging.LazyLogging
 import edu.ie3.simona.config.SimonaConfig.Simona.Output.Sink.InfluxDb1x
-import edu.ie3.simona.config.SimonaConfig.{BaseOutputConfig, RefSystemConfig}
+import edu.ie3.simona.config.SimonaConfig.{
+  BaseOutputConfig,
+  RefSystemConfig,
+  Simona,
+  TransformerControlGroup
+}
 import edu.ie3.simona.exceptions.InvalidConfigParameterException
 import edu.ie3.simona.io.result.ResultSinkType
 import edu.ie3.simona.model.participant.load.{LoadModelBehaviour, LoadReference}
@@ -128,6 +133,9 @@ case object ConfigFailFast extends LazyLogging {
 
     /* Check power flow resolution configuration */
     checkPowerFlowResolutionConfiguration(simonaConfig.simona.powerflow)
+
+    /* Check control scheme definitions */
+    simonaConfig.simona.control.foreach(checkControlSchemes)
   }
 
   /** Checks for valid sink configuration
@@ -395,15 +403,6 @@ case object ConfigFailFast extends LazyLogging {
           s"Provided refSystem is: $refSystem."
       )
 
-    voltLvls.foreach { voltLvl =>
-      {
-        if (!ConfigConventions.voltLvlRegex.matches(voltLvl))
-          throw new InvalidConfigParameterException(
-            s"The definition string for voltLvl '$voltLvl' does not comply with the definition {<id>, <rated voltage>}!"
-          )
-      }
-    }
-
     gridIds.foreach {
       case gridIdRange @ ConfigConventions.gridIdDotRange(from, to) =>
         rangeCheck(from.toInt, to.toInt, gridIdRange)
@@ -535,6 +534,50 @@ case object ConfigFailFast extends LazyLogging {
           s"the time resolution for power flow calculation is at least rounded to a full second!"
       )
     }
+  }
+
+  /** Check the validity of control scheme definitions
+    *
+    * @param control
+    *   Control scheme definitions
+    */
+  private def checkControlSchemes(control: Simona.Control): Unit = {
+    control.transformer.foreach(checkTransformerControl)
+  }
+
+  /** Check the suitability of transformer control group definition.
+    *
+    * One important check cannot be performed at this place, as input data is
+    * not available, yet: Do the measurements belong to a region, that can be
+    * influenced by the transformer?
+    *
+    * @param transformerControlGroup
+    *   Transformer control group definition
+    */
+  private def checkTransformerControl(
+      transformerControlGroup: TransformerControlGroup
+  ): Unit = transformerControlGroup match {
+    case TransformerControlGroup(measurements, transformers, vMax, vMin) =>
+      if (measurements.isEmpty)
+        throw new InvalidConfigParameterException(
+          "A transformer control group cannot have no measurements assigned."
+        )
+      if (transformers.isEmpty)
+        throw new InvalidConfigParameterException(
+          "A transformer control group cannot have no transformers assigned."
+        )
+      if (vMin < 0)
+        throw new InvalidConfigParameterException(
+          "The minimum permissible voltage magnitude of a transformer control group has to be positive."
+        )
+      if (vMax < 0)
+        throw new InvalidConfigParameterException(
+          "The maximum permissible voltage magnitude of a transformer control group has to be positive."
+        )
+      if (vMax < vMin)
+        throw new InvalidConfigParameterException(
+          "The minimum permissible voltage magnitude of a transformer control group must be smaller than the maximum permissible voltage magnitude."
+        )
   }
 
   /** Check the default config