+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Config possibility for transformer control groups
+
 ### Changed
 - Improving code readability in EvcsAgent by moving FreeLotsRequest to separate methods
 
-[Unreleased]: https://github.com/ie3-institute/simona
+[Unreleased]: https://github.com/ie3-institute/simona/compare/a14a093239f58fca9b2b974712686b33e5e5f939...HEAD